requirements.txt
Django==2.2.3
django-cors-middleware==1.3.1
django-rest-swagger==2.2.0
djangorestframework==3.9.4
djangorestframework-jwt==1.9.0
twilio
Pillow
django-fsm
authy
django-randompinfield
